+++ cardano-cli/src/Cardano/CLI/Shelley/Parsers.hs
@@ -739,7 +739,7 @@ pTransaction =
                         Nothing
                         "Filepath of auxiliary script(s)")
             <*> many pMetadataFile
-            <*> optional pProtocolParamsSourceSpec
+            <*> optional pProtocolParamsFile
             <*> optional pUpdateProposalFile
             <*> (OutputTxBodyOnly <$> pTxBodyFile Output <|> pCalculatePlutusScriptCost)
 
@@ -775,7 +775,7 @@ pTransaction =
                            Nothing
                            "Filepath of auxiliary script(s)")
                <*> many pMetadataFile
-               <*> optional pProtocolParamsSourceSpec
+               <*> optional pProtocolParamsFile
                <*> optional pUpdateProposalFile
                <*> pTxBodyFile Output
 
@@ -811,7 +811,7 @@ pTransaction =
     TxCalculateMinFee
       <$> pTxBodyFile Input
       <*> optional pNetworkId
-      <*> pProtocolParamsSourceSpec
+      <*> pProtocolParamsFile
       <*> pTxInCount
       <*> pTxOutCount
       <*> pTxShelleyWitnessCount
@@ -820,16 +820,8 @@ pTransaction =
   pTransactionCalculateMinReqUTxO :: Parser TransactionCmd
   pTransactionCalculateMinReqUTxO = TxCalculateMinRequiredUTxO
     <$> pCardanoEra
-    <*> pProtocolParamsSourceSpec
+    <*> pProtocolParamsFile
     <*> pTxOut
-
-  pProtocolParamsSourceSpec :: Parser ProtocolParamsSourceSpec
-  pProtocolParamsSourceSpec =
-    ParamsFromGenesis <$>
-      pGenesisFile
-        "[TESTING] The genesis file to take initial protocol parameters from.  For test clusters only, since the parameters are going to be obsolete for production clusters."
-    <|>
-    ParamsFromFile <$> pProtocolParamsFile
 
   pTxHashScriptData :: Parser TransactionCmd
   pTxHashScriptData = TxHashScriptData <$>
@@ -1469,7 +1461,6 @@ pAddressKeyType =
       )
   <|>
     pure AddressKeyShelley
-
 
 pProtocolParamsFile :: Parser ProtocolParamsFile
 pProtocolParamsFile =