+++ cardano-cli/ChangeLog.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Remove cardano-cli address build-script ([PR 4700](https://github.com/input-output-hk/cardano-node/pull/4700))
+- Remove support for reading protocol parameters from Shelley genesis file ([PR 5053](https://github.com/input-output-hk/cardano-node/pull/5053))
 
 ### Features